requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
diffusers>=0.21.0
accelerate>=0.20.0
datasets>=2.12.0
numpy>=1.24.0
pillow>=9.5.0
opencv-python>=4.8.0
matplotlib>=3.7.0
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0
einops>=0.6.0
xformers>=0.0.20
safetensors>=0.3.0
omegaconf>=2.3.0
hydra-core>=1.3.0 